% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{calc_marshalled_size}
\alias{calc_marshalled_size}
\title{Calculate the size of the R object when it is serialized}
\usage{
calc_marshalled_size(robj)
}
\arguments{
\item{robj}{R object}
}
\description{
Calculate the size of the R object when it is serialized
}
