% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{calc_serialized_size}
\alias{calc_serialized_size}
\title{Calculate the size of the R object when it is serialized}
\usage{
calc_serialized_size(robj)
}
\arguments{
\item{robj}{R object}
}
\value{
number of bytes needed to serialize this object
}
\description{
This function goes through all the motions of serializing an object, but
does nothing with the bytes other than to tally the total length.
}
\examples{
calc_serialized_size(mtcars)
}
