% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{init_smuggle}
\alias{init_smuggle}
\title{Init smuggle}
\usage{
init_smuggle(rc_xptr, wc_xptr)
}
\arguments{
\item{rc_xptr}{read connection (external pointer)}

\item{wc_xptr}{write connection (external pointer)}
}
\value{
None
}
\description{
Init smuggle
}
