% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{marshall}
\alias{marshall}
\title{Serialize an R object to a raw vector or connection.}
\usage{
marshall(robj, con = NULL)
}
\arguments{
\item{robj}{R object}

\item{con}{connection to write the data}
}
\value{
If \code{con} is a connection, return nothing.  Otherwise, return 
        a raw vector containing the serialized data.
}
\description{
Serialize an R object to a raw vector or connection.
}
\examples{
tmp <- tempfile()
marshall(mtcars, con = gzfile(tmp))
unmarshall(gzfile(tmp))
}
