% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{marshall_con}
\alias{marshall_con}
\title{Serialize an R object to a connection}
\usage{
marshall_con(robj, con = NULL)
}
\arguments{
\item{robj}{R object}

\item{con}{connection to write the data to}
}
\value{
\code{none}
}
\description{
Serialize an R object to a connection
}
\examples{
tmp <- tempfile()
marshall_con(mtcars, con = gzfile(tmp))
unmarshall_con(gzfile(tmp))
}
