% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{marshall_con_smuggle}
\alias{marshall_con_smuggle}
\title{Serialize an R object to a connection using an non-API function from R}
\usage{
marshall_con_smuggle(robj, con = NULL)
}
\arguments{
\item{robj}{R object}

\item{con}{connection to write the data to}
}
\value{
\code{none}
}
\description{
This approach will not pass \code{R CMD check} as the function `R_WriteConnection()`
is not part of the R API, but it is much much faster than the officially
blessed method of calling-back-into-R to write to a connection.
}
\examples{
tmp <- tempfile()
marshall_con_smuggle(mtcars, con = gzfile(tmp))
unmarshall_con_smuggle(gzfile(tmp))
}
