% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{marshall_minimize_malloc}
\alias{marshall_minimize_malloc}
\title{Serialize an R object to a raw vector, while minimising memory allocations}
\usage{
marshall_minimize_malloc(robj)
}
\arguments{
\item{robj}{R object}
}
\description{
This is a slightly more complex version of \code{marshall()} which
pre-calculates the size of the serialized representation.  This means there
only needs to be 1 memory allocation (not including those internal to R),
and gives a speed boost for medium-to-large objects
}
