% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{marshall_raw}
\alias{marshall_raw}
\title{Serialize an R object to a raw vector}
\usage{
marshall_raw(robj)
}
\arguments{
\item{robj}{R object}
}
\value{
a raw vector containing the serialized data.
}
\description{
Serialize an R object to a raw vector
}
\examples{
m <- marshall_raw(head(mtcars))
m
unmarshall_raw(m)
}
