% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{pack}
\alias{pack}
\alias{unpack}
\title{Serialize/unserialize an R object to/from a raw vector.}
\usage{
pack(robj)

unpack(raw_vec)
}
\arguments{
\item{robj}{R object}

\item{raw_vec}{R object serialized as a raw vector}
}
\description{
This is just a demonstration package showing how serialization may be done
directly in C, rather than using \code{base::serialize()}.
}
