% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{unmarshall}
\alias{unmarshall}
\title{Unserialize an R object from a raw vector or connection}
\usage{
unmarshall(x)
}
\arguments{
\item{x}{A raw vector containing serialized data or a connection}
}
\value{
Return the unserialized data
}
\description{
Unserialize an R object from a raw vector or connection
}
\examples{
tmp <- tempfile()
marshall(mtcars, con = gzfile(tmp))
unmarshall(gzfile(tmp))
}
