% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{unmarshall_con}
\alias{unmarshall_con}
\title{Unserialize an R object from a connection}
\usage{
unmarshall_con(con)
}
\arguments{
\item{con}{A connection from which to read serialized data}
}
\value{
Return the unserialized R object
}
\description{
Unserialize an R object from a connection
}
\examples{
tmp <- tempfile()
marshall_con(mtcars, con = gzfile(tmp))
unmarshall_con(gzfile(tmp))
}
