% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{unmarshall_con_illegal}
\alias{unmarshall_con_illegal}
\title{Unserialize an R object from a connection using an non-API function from R}
\usage{
unmarshall_con_illegal(con)
}
\arguments{
\item{con}{A connection from which to read serialized data}
}
\value{
Return the unserialized R object
}
\description{
This approach will not pass \code{R CMD check} as the function `R_ReadConnection()`
is not part of the R API, but it is much much faster than the officially
blessed method of calling-back-into-R to read from a connection.
}
\examples{
tmp <- tempfile()
marshall_con_illegal(mtcars, con = gzfile(tmp))
unmarshall_con_illegal(gzfile(tmp))
}
